+++ internal/configs/main.go
@@ -25,6 +25,24 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
+// Metric is a metric definition for the log config
+type Metric struct {
+	Matcher    *regexp.Regexp
+	MatchParts []string
+	Namer      *template.Template
+	ValueKey   string
+	Match      string `json:"match" yaml:"match" toml:"match"`
+	Name       string `json:"name" yaml:"name" toml:"name"`
+	Type       string `json:"type" yaml:"type" toml:"type"`
+}
+
+// Config defines a log to watch
+type Config struct {
+	ID      string    `json:"id" yaml:"id" toml:"id"`
+	LogFile string    `json:"log_file" yaml:"log_file" toml:"log_file"`
+	Metrics []*Metric `json:"metrics" yaml:"metrics" toml:"metrics"`
+}
+
 // Load reads the log configurations from log config directory
 func Load() ([]*Config, error) {
 	logger := log.With().Str("pkg", "configs").Logger()
